requirements.txt
flask>=3.0
werkzeug>=3.0
pillow>=10.0
numpy>=1.26
tensorflow>=2.15
flask-bcrypt>=1.0
gradio-client>=0.6.0
